+++ getset.go
@@ -15,68 +15,75 @@ func (c *Config) CountField(name string) (int, error) {
 	return -1, raise(ErrMissing)
 }
 
-func (c *Config) Bool(name string, idx int) (bool, error) {
-	v, err := c.getField(name, idx)
+func (c *Config) Bool(name string, idx int, opts ...Option) (bool, error) {
+	v, err := c.getField(name, idx, opts)
 	if err != nil {
 		return false, err
 	}
 	return v.toBool()
 }
 
-func (c *Config) String(name string, idx int) (string, error) {
-	v, err := c.getField(name, idx)
+func (c *Config) String(name string, idx int, opts ...Option) (string, error) {
+	v, err := c.getField(name, idx, opts)
 	if err != nil {
 		return "", err
 	}
 	return v.toString()
 }
 
-func (c *Config) Int(name string, idx int) (int64, error) {
-	v, err := c.getField(name, idx)
+func (c *Config) Int(name string, idx int, opts ...Option) (int64, error) {
+	v, err := c.getField(name, idx, opts)
 	if err != nil {
 		return 0, err
 	}
 	return v.toInt()
 }
 
-func (c *Config) Float(name string, idx int) (float64, error) {
-	v, err := c.getField(name, idx)
+func (c *Config) Float(name string, idx int, opts ...Option) (float64, error) {
+	v, err := c.getField(name, idx, opts)
 	if err != nil {
 		return 0, err
 	}
 	return v.toFloat()
 }
 
-func (c *Config) Child(name string, idx int) (*Config, error) {
-	v, err := c.getField(name, idx)
+func (c *Config) Child(name string, idx int, opts ...Option) (*Config, error) {
+	v, err := c.getField(name, idx, opts)
 	if err != nil {
 		return nil, err
 	}
 	return v.toConfig()
 }
 
-func (c *Config) SetBool(name string, idx int, value bool) {
-	c.setField(name, idx, &cfgBool{b: value})
+func (c *Config) SetBool(name string, idx int, value bool, opts ...Option) error {
+	return c.setField(name, idx, &cfgBool{b: value}, opts)
 }
 
-func (c *Config) SetInt(name string, idx int, value int64) {
-	c.setField(name, idx, &cfgInt{i: value})
+func (c *Config) SetInt(name string, idx int, value int64, opts ...Option) error {
+	return c.setField(name, idx, &cfgInt{i: value}, opts)
 }
 
-func (c *Config) SetFloat(name string, idx int, value float64) {
-	c.setField(name, idx, &cfgFloat{f: value})
+func (c *Config) SetFloat(name string, idx int, value float64, opts ...Option) error {
+	return c.setField(name, idx, &cfgFloat{f: value}, opts)
 }
 
-func (c *Config) SetString(name string, idx int, value string) {
-	c.setField(name, idx, &cfgString{s: value})
+func (c *Config) SetString(name string, idx int, value string, opts ...Option) error {
+	return c.setField(name, idx, &cfgString{s: value}, opts)
 }
 
-func (c *Config) SetChild(name string, idx int, value *Config) {
-	c.setField(name, idx, cfgSub{c: value})
+func (c *Config) SetChild(name string, idx int, value *Config, opts ...Option) error {
+	return c.setField(name, idx, cfgSub{c: value}, opts)
 }
 
-func (c *Config) getField(name string, idx int) (value, error) {
-	v, ok := c.fields.fields[name]
+func (c *Config) getField(name string, idx int, options []Option) (value, error) {
+	opts := makeOptions(options)
+
+	cfg, field, err := reifyCfgPath(c, opts, name)
+	if err != nil {
+		return nil, err
+	}
+
+	v, ok := cfg.fields.fields[field]
 	if !ok {
 		return nil, raise(ErrMissing)
 	}
@@ -95,14 +102,20 @@ func (c *Config) getField(name string, idx int) (value, error) {
 	return v, nil
 }
 
-func (c *Config) setField(name string, idx int, v value) {
+func (c *Config) setField(name string, idx int, v value, options []Option) error {
+	opts := makeOptions(options)
 	ctx := context{
 		parent: cfgSub{c},
 		field:  name,
 	}
 	orig := v
 
-	old, ok := c.fields.fields[name]
+	cfg, field, err := normalizeCfgPath(c, opts, name)
+	if err != nil {
+		return err
+	}
+
+	old, ok := cfg.fields.fields[field]
 	if !ok {
 		if idx > 0 {
 			slice := &cfgArray{
@@ -137,5 +150,7 @@ func (c *Config) setField(name string, idx int, v value) {
 		ctx.field = fmt.Sprintf("%v", idx)
 	}
 	orig.SetContext(ctx)
-	c.fields.fields[name] = v
+	cfg.fields.fields[field] = v
+
+	return nil
 }