+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
+- Add options and PathSep support to low level option setters/getters
 - Added support for _rebranding_ `*ucfg.Config` via `type MyConfig ucfg.Config` using
   casts between pointer types in Unpack and Merge.
 - Introduced CHANGELOG.md for documenting changes to ucfg.